+++ Extra/XCode Templates/VCore (SwiftUI)/ContentView (MV).xctemplate/___FILEBASENAME___Parameters.swift
@@ -1,0 +1,16 @@
+//  ___FILEHEADER___
+
+import Foundation
+
+// MARK: - ___VARIABLE_productName___ Parameters
+struct ___VARIABLE_productName___Parameters: Hashable {
+    // MARK: Properties
+    // ...
+    
+    // MARK: Mock
+#if DEBUG
+    static var mock: Self {
+        .init()
+    }
+#endif
+}